python/requirements.txt
requests>=2.32.0
python-dotenv>=1.0.1
colorama>=0.4.6
